+++ flink-1.17-connector-bigquery/flink-connector-bigquery/src/main/java/com/google/cloud/flink/bigquery/sink/BigQuerySinkConfig.java
@@ -16,11 +16,24 @@
 
 package com.google.cloud.flink.bigquery.sink;
 
+import org.apache.flink.annotation.Internal;
+import org.apache.flink.api.common.restartstrategy.RestartStrategies.ExponentialDelayRestartStrategyConfiguration;
+import org.apache.flink.api.common.restartstrategy.RestartStrategies.FailureRateRestartStrategyConfiguration;
+import org.apache.flink.api.common.restartstrategy.RestartStrategies.FixedDelayRestartStrategyConfiguration;
+import org.apache.flink.api.common.restartstrategy.RestartStrategies.NoRestartStrategyConfiguration;
+import org.apache.flink.api.common.restartstrategy.RestartStrategies.RestartStrategyConfiguration;
 import org.apache.flink.connector.base.DeliveryGuarantee;
+import org.apache.flink.streaming.api.environment.StreamExecutionEnvironment;
+import org.apache.flink.table.types.logical.LogicalType;
 
 import com.google.cloud.flink.bigquery.common.config.BigQueryConnectOptions;
 import com.google.cloud.flink.bigquery.sink.serializer.BigQueryProtoSerializer;
 import com.google.cloud.flink.bigquery.sink.serializer.BigQuerySchemaProvider;
+import com.google.cloud.flink.bigquery.sink.serializer.BigQuerySchemaProviderImpl;
+import com.google.cloud.flink.bigquery.sink.serializer.BigQueryTableSchemaProvider;
+import com.google.cloud.flink.bigquery.sink.serializer.RowDataToProtoSerializer;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 import java.util.Objects;
 
@@ -30,6 +43,11 @@ import java.util.Objects;
  * <p>Uses static inner builder to initialize new instances.
  */
 public class BigQuerySinkConfig {
+
+    private static final Logger LOG = LoggerFactory.getLogger(BigQuerySink.class);
+    private static final long MILLISECONDS_PER_SECOND = 1000L;
+    private static final long MILLISECONDS_PER_MINUTE = 60L * 1000L;
+    private static final long MILLISECONDS_PER_HOUR = 60L * 60L * 1000L;
 
     private final BigQueryConnectOptions connectOptions;
     private final DeliveryGuarantee deliveryGuarantee;
@@ -59,7 +77,7 @@ public class BigQuerySinkConfig {
         }
         BigQuerySinkConfig object = (BigQuerySinkConfig) obj;
         if (this.getConnectOptions() == object.getConnectOptions()
-                && (this.getSerializer() == object.getSerializer())
+                && (this.getSerializer().getClass() == object.getSerializer().getClass())
                 && (this.getDeliveryGuarantee() == object.getDeliveryGuarantee())) {
             BigQuerySchemaProvider thisSchemaProvider = this.getSchemaProvider();
             BigQuerySchemaProvider objSchemaProvider = object.getSchemaProvider();
@@ -102,6 +120,7 @@ public class BigQuerySinkConfig {
         private DeliveryGuarantee deliveryGuarantee;
         private BigQuerySchemaProvider schemaProvider;
         private BigQueryProtoSerializer serializer;
+        private StreamExecutionEnvironment env;
 
         public Builder connectOptions(BigQueryConnectOptions connectOptions) {
             this.connectOptions = connectOptions;
@@ -123,9 +142,141 @@ public class BigQuerySinkConfig {
             return this;
         }
 
+        public Builder streamExecutionEnvironment(
+                StreamExecutionEnvironment streamExecutionEnvironment) {
+            this.env = streamExecutionEnvironment;
+            return this;
+        }
+
         public BigQuerySinkConfig build() {
+            if (deliveryGuarantee == DeliveryGuarantee.EXACTLY_ONCE) {
+                validateStreamExecutionEnvironment(env);
+            }
             return new BigQuerySinkConfig(
                     connectOptions, deliveryGuarantee, schemaProvider, serializer);
+        }
+    }
+
+    // DO NOT USE!
+    // This method is used internally to create sink config for the Table API integration.
+    // Note that the serializer is hard coded for Flink Table's RowData.
+    @Internal
+    public static BigQuerySinkConfig forTable(
+            BigQueryConnectOptions connectOptions,
+            DeliveryGuarantee deliveryGuarantee,
+            LogicalType logicalType) {
+        return new BigQuerySinkConfig(
+                connectOptions,
+                deliveryGuarantee,
+                new BigQuerySchemaProviderImpl(
+                        BigQueryTableSchemaProvider.getAvroSchemaFromLogicalSchema(logicalType)),
+                new RowDataToProtoSerializer());
+    }
+
+    public static void validateStreamExecutionEnvironment(StreamExecutionEnvironment env) {
+        if (env == null) {
+            throw new IllegalArgumentException(
+                    "Expected StreamExecutionEnvironment, found null."
+                            + " Please provide the StreamExecutionEnvironment used in Flink job.");
+        }
+        validateRestartStrategy(env.getRestartStrategy());
+    }
+
+    private static void validateRestartStrategy(RestartStrategyConfiguration restartStrategy) {
+        if (restartStrategy == null) {
+            throw new IllegalArgumentException(
+                    "Could not read RestartStrategyConfiguration from StreamExecutionEnvironment."
+                            + " Please provide the StreamExecutionEnvironment used in Flink job and"
+                            + " set a restart strategy.");
+        }
+        // Restart configurations are mostly being checked against Flink defaults for optimum
+        // interactions with external systems, primarily BigQuery storage write APIs.
+        // Keep in mind, maximum 10,000 CreateWriteStream calls are allowed by BigQuery per hour per
+        // project per region.
+        if (restartStrategy instanceof FixedDelayRestartStrategyConfiguration) {
+            FixedDelayRestartStrategyConfiguration strategy =
+                    (FixedDelayRestartStrategyConfiguration) restartStrategy;
+            if (strategy.getDelayBetweenAttemptsInterval().toMilliseconds()
+                            < MILLISECONDS_PER_SECOND
+                    || strategy.getRestartAttempts() > 10) {
+                LOG.error(
+                        "Invalid FixedDelayRestartStrategyConfiguration: found restart delay {},"
+                                + " milliseconds, and {} restart attempts. Should be used with"
+                                + " restart delay at least 1 second, and at most 10 restart"
+                                + " attempts.",
+                        strategy.getDelayBetweenAttemptsInterval().toMilliseconds(),
+                        strategy.getRestartAttempts());
+                throw new IllegalArgumentException(
+                        "Invalid restart strategy: FixedDelayRestartStrategyConfiguration should"
+                                + " be used with at least restart delay 1 second, and at most 10"
+                                + " restart attempts.");
+            }
+        } else if (restartStrategy instanceof ExponentialDelayRestartStrategyConfiguration) {
+            ExponentialDelayRestartStrategyConfiguration strategy =
+                    (ExponentialDelayRestartStrategyConfiguration) restartStrategy;
+            if (strategy.getBackoffMultiplier() < 2.0
+                    || strategy.getInitialBackoff().toMilliseconds() < MILLISECONDS_PER_SECOND
+                    || strategy.getMaxBackoff().toMilliseconds() < (5L * MILLISECONDS_PER_MINUTE)
+                    || strategy.getResetBackoffThreshold().toMilliseconds()
+                            < MILLISECONDS_PER_HOUR) {
+                LOG.error(
+                        "Invalid ExponentialDelayRestartStrategyConfiguration: found backoff"
+                                + " multiplier {}, initial backoff {} milliseconds, maximum backoff"
+                                + " {} milliseconds, and reset threshold {} milliseconds. Should be"
+                                + " used with backoff multiplier at least 2, initial backoff at"
+                                + " least 1 second, maximum backoff at least 5 minutes, and reset"
+                                + " threshold at least 1 hour",
+                        strategy.getBackoffMultiplier(),
+                        strategy.getInitialBackoff().toMilliseconds(),
+                        strategy.getMaxBackoff().toMilliseconds(),
+                        strategy.getResetBackoffThreshold().toMilliseconds());
+                throw new IllegalArgumentException(
+                        "Invalid restart strategy: ExponentialDelayRestartStrategyConfiguration"
+                                + " should be used with backoff multiplier at least 2, initial"
+                                + " backoff at least 1 second, maximum backoff at-least 5 minutes,"
+                                + " and reset threshold at least 1 hour");
+            }
+        } else if (restartStrategy instanceof FailureRateRestartStrategyConfiguration) {
+            FailureRateRestartStrategyConfiguration strategy =
+                    (FailureRateRestartStrategyConfiguration) restartStrategy;
+            double failureIntervalInMinutes =
+                    ((double) strategy.getFailureInterval().toMilliseconds())
+                            / MILLISECONDS_PER_MINUTE;
+            double allowedFailuresPerMinute =
+                    ((double) strategy.getMaxFailureRate()) / failureIntervalInMinutes;
+            if (strategy.getDelayBetweenAttemptsInterval().toMilliseconds()
+                            < MILLISECONDS_PER_SECOND
+                    || allowedFailuresPerMinute > 1.0) {
+                LOG.error(
+                        "Invalid FailureRateRestartStrategyConfiguration: found restart delay {}"
+                                + " milliseconds, and allowed failure rate {} per minute. Should be"
+                                + " used with restart delay at least 1 second, and allowed failure"
+                                + " rate at most 1 per minute.",
+                        strategy.getDelayBetweenAttemptsInterval().toMilliseconds(),
+                        allowedFailuresPerMinute);
+                throw new IllegalArgumentException(
+                        "Invalid restart strategy: FailureRateRestartStrategyConfiguration should"
+                                + " be used with restart delay at least 1 second, and allowed"
+                                + " failure rate at most 1 per minute.");
+            }
+        } else if (restartStrategy instanceof NoRestartStrategyConfiguration) {
+            // This is fine!
+            LOG.debug("Found NoRestartStrategyConfiguration. No validation needed.");
+        } else {
+            // Can be either FallbackRestartStrategyConfiguration, or a custom implementation. We
+            // don't know how to verify these two possibilities, so we let them pass. If you,
+            // worthy reader have a suggestion for validating these scenarios, then kindly share
+            // with us at
+            // https://github.com/GoogleCloudDataproc/flink-bigquery-connector/issues/new
+            // Thanks!
+            LOG.warn(
+                    "Cannot validate RestartStrategyConfiguration in StreamExecutionEnvironment."
+                            + " We recommend explicitly setting the restart strategy as one of the"
+                            + " following:"
+                            + " FixedDelayRestartStrategyConfiguration,"
+                            + " ExponentialDelayRestartStrategyConfiguration,"
+                            + " FailureRateRestartStrategyConfiguration or"
+                            + " NoRestartStrategyConfiguration");
         }
     }
 }