+++ flink-1.17-connector-bigquery/flink-connector-bigquery/src/main/java/com/google/cloud/flink/bigquery/sink/writer/BaseWriter.java
@@ -18,6 +18,8 @@ package com.google.cloud.flink.bigquery.sink.writer;
 
 import org.apache.flink.annotation.Internal;
 import org.apache.flink.api.connector.sink2.SinkWriter;
+import org.apache.flink.metrics.Counter;
+import org.apache.flink.metrics.groups.SinkWriterMetricGroup;
 
 import com.google.api.core.ApiFuture;
 import com.google.cloud.bigquery.storage.v1.AppendRowsResponse;
@@ -34,6 +36,7 @@ import com.google.cloud.flink.bigquery.sink.exceptions.BigQuerySerializationExce
 import com.google.cloud.flink.bigquery.sink.serializer.BigQueryProtoSerializer;
 import com.google.cloud.flink.bigquery.sink.serializer.BigQuerySchemaProvider;
 import com.google.protobuf.ByteString;
+import com.google.protobuf.Descriptors.Descriptor;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -64,7 +67,6 @@ import java.util.Queue;
 abstract class BaseWriter<IN> implements SinkWriter<IN> {
 
     protected final Logger logger = LoggerFactory.getLogger(getClass());
-
     // Multiply 0.95 to keep a buffer from exceeding payload limits.
     private static final long MAX_APPEND_REQUEST_BYTES =
             (long) (StreamWriter.getApiMaxRequestBytes() * 0.95);
@@ -90,6 +92,10 @@ abstract class BaseWriter<IN> implements SinkWriter<IN> {
     // the records in a stream get committed to the table. Hence, records written to BigQuery
     // table is equal to this "totalRecordsWritten" only upon checkpoint completion.
     long totalRecordsWritten;
+    // Counters for metric reporting
+    Counter numberOfRecordsWrittenToBigQuery;
+    Counter numberOfRecordsSeenByWriter;
+    Counter numberOfRecordsSeenByWriterSinceCheckpoint;
 
     BaseWriter(
             int subtaskId,
@@ -254,6 +260,21 @@ abstract class BaseWriter<IN> implements SinkWriter<IN> {
                         "AppendRows request failed in subtask %d\n%s", subtaskId, errorMessage));
         throw new BigQueryConnectorException(
                 String.format("Error while writing to BigQuery\n%s", errorMessage));
+    }
+
+    /**
+     * Method to initialize flink metrics which are common across at-least-once and exactly-once
+     * approach.
+     *
+     * @param sinkWriterMetricGroup Metric Group to register the counters.
+     */
+    void initializeMetrics(SinkWriterMetricGroup sinkWriterMetricGroup) {
+        numberOfRecordsSeenByWriter = sinkWriterMetricGroup.counter("numberOfRecordsSeenByWriter");
+        // Count of records which are successfully appended to BQ.
+        numberOfRecordsWrittenToBigQuery =
+                sinkWriterMetricGroup.counter("numberOfRecordsWrittenToBigQuery");
+        numberOfRecordsSeenByWriterSinceCheckpoint =
+                sinkWriterMetricGroup.counter("numberOfRecordsSeenByWriterSinceCheckpoint");
     }
 
     static class AppendInfo {